+++ public/pages/workflow_detail/components/export_modal.tsx
@@ -3,76 +3,48 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import React, { useEffect, useState } from 'react';
-import yaml from 'js-yaml';
-import { isEmpty, toLower } from 'lodash';
+import React, { useState } from 'react';
 import {
-  EuiCodeBlock,
   EuiFlexGroup,
-  EuiFlexItem,
-  EuiText,
-  EuiLink,
   EuiModal,
   EuiModalHeader,
   EuiModalHeaderTitle,
   EuiModalBody,
   EuiModalFooter,
   EuiSmallButtonEmpty,
-  EuiSmallButtonGroup,
-  EuiCallOut,
+  EuiTabs,
+  EuiTab,
+  EuiFlexItem,
 } from '@elastic/eui';
 import {
-  CREATE_WORKFLOW_LINK,
   Workflow,
-  customStringify,
+  WORKFLOW_TYPE,
   getCharacterLimitedString,
 } from '../../../../common';
-import { reduceToTemplate } from '../../../utils';
 import '../../../global-styles.scss';
+import { ExportTemplateContent } from './export_template_content';
+import { AgenticSearchApplicationContent } from './agentic_search_application_content';
 
 interface ExportModalProps {
   workflow?: Workflow;
   setIsExportModalOpen(isOpen: boolean): void;
 }
 
-enum EXPORT_OPTION {
-  JSON = 'JSON',
-  YAML = 'YAML',
+enum EXPORT_TAB {
+  TEMPLATE = 'TEMPLATE',
+  APPLICATION = 'APPLICATION',
 }
 
 /**
  * Modal containing all of the export options
  */
 export function ExportModal(props: ExportModalProps) {
-  // format type state
-  const [selectedOption, setSelectedOption] = useState<EXPORT_OPTION>(
-    EXPORT_OPTION.JSON
+  const isAgenticSearchType =
+    props.workflow?.ui_metadata?.type === WORKFLOW_TYPE.AGENTIC_SEARCH;
+
+  const [selectedTab, setSelectedTab] = useState<EXPORT_TAB>(
+    isAgenticSearchType ? EXPORT_TAB.APPLICATION : EXPORT_TAB.TEMPLATE
   );
-
-  // formatted string state
-  const [formattedConfig, setFormattedConfig] = useState<string>('');
-  useEffect(() => {
-    if (props.workflow) {
-      const workflowTemplate = reduceToTemplate(props.workflow);
-      if (selectedOption === EXPORT_OPTION.JSON) {
-        setFormattedConfig(customStringify(workflowTemplate));
-      } else if (selectedOption === EXPORT_OPTION.YAML) {
-        setFormattedConfig(yaml.dump(workflowTemplate));
-      }
-    }
-  }, [props.workflow, selectedOption]);
-
-  // client-side file to be downloaded if the user so chooses. Generate a file
-  // and its corresponding URL.
-  const [formattedConfigHref, setFormattedConfigHref] = useState<string>('');
-  useEffect(() => {
-    if (!isEmpty(formattedConfig)) {
-      const formattedConfigFile = new Blob([formattedConfig], {
-        type: `text/${toLower(selectedOption)}`,
-      });
-      setFormattedConfigHref(URL.createObjectURL(formattedConfigFile));
-    }
-  }, [formattedConfig]);
 
   return (
     <EuiModal
@@ -89,75 +61,39 @@ export function ExportModal(props: ExportModalProps) {
         </EuiModalHeaderTitle>
       </EuiModalHeader>
       <EuiModalBody>
-        <EuiFlexGroup direction="column">
-          {isEmpty(props.workflow?.workflows) && (
+        {isAgenticSearchType ? (
+          <EuiFlexGroup
+            direction="column"
+            gutterSize="s"
+            style={{ marginTop: '-16px' }}
+          >
             <EuiFlexItem grow={false}>
-              <EuiCallOut color="warning" size="s" iconType={'alert'}>
-                This workflow will provision no resources. You may still export
-                to save your configuration.
-              </EuiCallOut>
-            </EuiFlexItem>
-          )}
-          <EuiFlexItem grow={false}>
-            <EuiText size="s">
-              {`To build identical resources in other environments, create and provision a workflow following the below template.`}{' '}
-              <EuiLink href={CREATE_WORKFLOW_LINK} target="_blank">
-                Learn more
-              </EuiLink>
-            </EuiText>
-            <EuiText
-              size="s"
-              color="subdued"
-            >{`Note: Certain resource IDs in the template, such as model IDs, may be specific to a cluster and not function properly 
-            in other clusters. Make sure to update these values before provisioning the workflow in a new cluster.`}</EuiText>
-          </EuiFlexItem>
-          <EuiFlexItem>
-            <EuiFlexGroup direction="row" justifyContent="spaceBetween">
-              <EuiFlexItem grow={false}>
-                <EuiSmallButtonGroup
-                  legend="Choose how to view your workflow"
-                  options={[
-                    {
-                      id: EXPORT_OPTION.JSON,
-                      label: EXPORT_OPTION.JSON,
-                    },
-                    {
-                      id: EXPORT_OPTION.YAML,
-                      label: EXPORT_OPTION.YAML,
-                    },
-                  ]}
-                  idSelected={selectedOption}
-                  onChange={(id) => setSelectedOption(id as EXPORT_OPTION)}
-                  data-testid="exportDataToggleButtonGroup"
-                />
-              </EuiFlexItem>
-              <EuiFlexItem grow={false}>
-                <EuiSmallButtonEmpty
-                  iconType="download"
-                  iconSide="right"
-                  href={formattedConfigHref}
-                  download={`${props.workflow?.name}.${toLower(
-                    selectedOption
-                  )}`}
-                  onClick={() => {}}
+              <EuiTabs size="s" data-testid="agenticSearchTabs">
+                <EuiTab
+                  onClick={() => setSelectedTab(EXPORT_TAB.APPLICATION)}
+                  isSelected={selectedTab === EXPORT_TAB.APPLICATION}
                 >
-                  {`Download ${selectedOption} file`}
-                </EuiSmallButtonEmpty>
-              </EuiFlexItem>
-            </EuiFlexGroup>
-          </EuiFlexItem>
-          {props.workflow !== undefined && (
-            <EuiFlexItem grow={false}>
-              <EuiCodeBlock
-                language={toLower(selectedOption)}
-                fontSize="m"
-                isCopyable={true}
-              >
-                {formattedConfig}
-              </EuiCodeBlock>
+                  Use in your application
+                </EuiTab>
+                <EuiTab
+                  onClick={() => setSelectedTab(EXPORT_TAB.TEMPLATE)}
+                  isSelected={selectedTab === EXPORT_TAB.TEMPLATE}
+                >
+                  Template
+                </EuiTab>
+              </EuiTabs>
             </EuiFlexItem>
-          )}
-        </EuiFlexGroup>
+            <EuiFlexItem grow={false}>
+              {selectedTab === EXPORT_TAB.TEMPLATE ? (
+                <ExportTemplateContent workflow={props.workflow} />
+              ) : (
+                <AgenticSearchApplicationContent workflow={props.workflow} />
+              )}
+            </EuiFlexItem>
+          </EuiFlexGroup>
+        ) : (
+          <ExportTemplateContent workflow={props.workflow} />
+        )}
       </EuiModalBody>
       <EuiModalFooter>
         <EuiSmallButtonEmpty